requirements.txt
#
# This file is autogenerated by pip-compile with python 3.7
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
anyio==3.5.0
    # via jupyter-server
argon2-cffi==21.3.0
    # via
    #   jupyter-server
    #   notebook
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
attrs==21.4.0
    # via
    #   fiona
    #   jsonschema
babel==2.9.1
    # via jupyterlab-server
backcall==0.2.0
    # via ipython
beautifulsoup4==4.10.0
    # via nbconvert
black==22.3.0
    # via -r requirements.in
bleach==4.1.0
    # via nbconvert
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via
    #   fiona
    #   httplib2shim
    #   pyproj
    #   requests
cffi==1.15.0
    # via argon2-cffi-bindings
charset-normalizer==2.0.12
    # via requests
click==8.1.0
    # via
    #   -r requirements.in
    #   black
    #   click-plugins
    #   cligj
    #   fiona
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
debugpy==1.6.0
    # via ipykernel
decorator==5.1.1
    # via ipython
defusedxml==0.7.1
    # via nbconvert
earthengine-api==0.1.303
    # via -r requirements.in
entrypoints==0.4
    # via
    #   jupyter-client
    #   jupyterlab-server
    #   nbconvert
fiona==1.8.21
    # via geopandas
future==0.18.2
    # via earthengine-api
geopandas==0.10.2
    # via -r requirements.in
google-api-core==2.7.1
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==1.12.11
    # via earthengine-api
google-auth==2.6.2
    # via
    #   earthengine-api
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.1.0
    # via
    #   earthengine-api
    #   google-api-python-client
google-cloud-core==2.2.3
    # via google-cloud-storage
google-cloud-storage==2.2.1
    # via earthengine-api
google-crc32c==1.3.0
    # via google-resumable-media
google-resumable-media==2.3.2
    # via google-cloud-storage
googleapis-common-protos==1.56.0
    # via google-api-core
httplib2==0.20.4
    # via
    #   earthengine-api
    #   google-api-python-client
    #   google-auth-httplib2
    #   httplib2shim
httplib2shim==0.0.3
    # via earthengine-api
idna==3.3
    # via
    #   anyio
    #   requests
importlib-metadata==4.11.3
    # via
    #   click
    #   jsonschema
importlib-resources==5.6.0
    # via jsonschema
ipykernel==6.10.0
    # via
    #   ipywidgets
    #   jupyter
    #   jupyter-console
    #   notebook
    #   qtconsole
ipython==7.32.0
    # via
    #   ipykernel
    #   ipywidgets
    #   jupyter-console
    #   jupyterlab
ipython-genutils==0.2.0
    # via
    #   ipywidgets
    #   notebook
    #   qtconsole
ipywidgets==7.7.0
    # via jupyter
jedi==0.18.1
    # via ipython
jinja2==3.1.1
    # via
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   nbconvert
    #   notebook
joblib==1.1.0
    # via -r requirements.in
json5==0.9.6
    # via jupyterlab-server
jsonschema==4.4.0
    # via
    #   jupyterlab-server
    #   nbformat
jupyter==1.0.0
    # via -r requirements.in
jupyter-client==7.2.0
    # via
    #   ipykernel
    #   jupyter-console
    #   jupyter-server
    #   nbclient
    #   notebook
    #   qtconsole
jupyter-console==6.4.3
    # via jupyter
jupyter-core==4.9.2
    # via
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
jupyter-server==1.16.0
    # via
    #   jupyterlab
    #   jupyterlab-server
    #   nbclassic
    #   notebook-shim
jupyterlab==3.3.2
    # via -r requirements.in
jupyterlab-pygments==0.1.2
    # via nbconvert
jupyterlab-server==2.12.0
    # via jupyterlab
jupyterlab-widgets==1.1.0
    # via ipywidgets
loguru==0.6.0
    # via -r requirements.in
markupsafe==2.1.1
    # via
    #   jinja2
    #   nbconvert
matplotlib-inline==0.1.3
    # via
    #   ipykernel
    #   ipython
mistune==0.8.4
    # via nbconvert
munch==2.5.0
    # via fiona
mypy-extensions==0.4.3
    # via black
nbclassic==0.3.7
    # via jupyterlab
nbclient==0.5.13
    # via nbconvert
nbconvert==6.4.5
    # via
    #   jupyter
    #   jupyter-server
    #   notebook
nbformat==5.2.0
    # via
    #   ipywidgets
    #   jupyter-server
    #   nbclient
    #   nbconvert
    #   notebook
nest-asyncio==1.5.4
    # via
    #   ipykernel
    #   jupyter-client
    #   nbclient
    #   notebook
notebook==6.4.10
    # via
    #   jupyter
    #   nbclassic
    #   widgetsnbextension
notebook-shim==0.1.0
    # via nbclassic
numpy==1.21.5
    # via
    #   -r requirements.in
    #   pandas
packaging==21.3
    # via
    #   bleach
    #   jupyter-server
    #   jupyterlab
    #   jupyterlab-server
    #   qtpy
pandas==1.3.5
    # via
    #   -r requirements.in
    #   geopandas
pandocfilters==1.5.0
    # via nbconvert
parso==0.8.3
    # via jedi
pathspec==0.9.0
    # via black
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
platformdirs==2.5.1
    # via black
prometheus-client==0.13.1
    # via
    #   jupyter-server
    #   notebook
prompt-toolkit==3.0.28
    # via
    #   ipython
    #   jupyter-console
protobuf==3.19.4
    # via
    #   google-api-core
    #   google-cloud-storage
    #   googleapis-common-protos
psutil==5.9.0
    # via ipykernel
ptyprocess==0.7.0
    # via
    #   pexpect
    #   terminado
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via -r requirements.in
pygments==2.11.2
    # via
    #   ipython
    #   jupyter-console
    #   jupyterlab-pygments
    #   nbconvert
    #   qtconsole
pyparsing==3.0.7
    # via
    #   httplib2
    #   packaging
pyproj==3.2.1
    # via geopandas
pyrsistent==0.18.1
    # via jsonschema
python-dateutil==2.8.2
    # via
    #   jupyter-client
    #   pandas
pytz==2022.1
    # via
    #   babel
    #   pandas
pyyaml==6.0
    # via -r requirements.in
pyzmq==22.3.0
    # via
    #   jupyter-client
    #   jupyter-server
    #   notebook
    #   qtconsole
qtconsole==5.3.0
    # via jupyter
qtpy==2.0.1
    # via qtconsole
requests==2.27.1
    # via
    #   google-api-core
    #   google-cloud-storage
    #   jupyterlab-server
rsa==4.8
    # via google-auth
send2trash==1.8.0
    # via
    #   jupyter-server
    #   notebook
shapely==1.8.1.post1
    # via geopandas
six==1.16.0
    # via
    #   bleach
    #   earthengine-api
    #   fiona
    #   google-api-python-client
    #   google-auth
    #   google-auth-httplib2
    #   httplib2shim
    #   munch
    #   python-dateutil
sniffio==1.2.0
    # via anyio
soupsieve==2.3.1
    # via beautifulsoup4
terminado==0.13.3
    # via
    #   jupyter-server
    #   notebook
testpath==0.6.0
    # via nbconvert
tomli==2.0.1
    # via black
tornado==6.1
    # via
    #   ipykernel
    #   jupyter-client
    #   jupyter-server
    #   jupyterlab
    #   notebook
    #   terminado
traitlets==5.1.1
    # via
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   jupyter-server
    #   matplotlib-inline
    #   nbclient
    #   nbconvert
    #   nbformat
    #   notebook
    #   qtconsole
typed-ast==1.5.2
    # via black
typing-extensions==4.1.1
    # via
    #   anyio
    #   argon2-cffi
    #   black
    #   importlib-metadata
    #   jsonschema
    #   pydantic
uritemplate==3.0.1
    # via google-api-python-client
urllib3==1.26.9
    # via
    #   httplib2shim
    #   requests
wcwidth==0.2.5
    # via prompt-toolkit
webencodings==0.5.1
    # via bleach
websocket-client==1.3.2
    # via jupyter-server
widgetsnbextension==3.6.0
    # via ipywidgets
zipp==3.7.0
    # via
    #   importlib-metadata
    #   importlib-resources

# The following packages are considered to be unsafe in a requirements file:
# setuptools
